producer-service/requirements.txt
google-cloud-pubsub
google-cloud-storage
confluent-kafka